Centrality Algorithms

Centrality Algorithms

EvolvingGraphs.Centrality.katz — Function.

katz(g [, α = 0.3, sorted = true])

Computes the broadcast centrality vector of an evolving graph g.

Input:

 `g`: an evolving graph
 `α`: (= 0.3 default) a scalar the controls the influence of long walks.

source

katz(g, α, β [, mode = :broadcast])

Computes the Katz centrality of an evolving graph g.

Input:

 `g`: an evolving graph
 `α`: a scalar that controls the influence of long walks
 `β`: a scalar that controls the influence of walks happened long time ago.
      `mode`: `mode = :broadcast` (default) generates the broadcast centrality
      vector; `mode = :receive` generates the receving centrality vector;
      `mode = :matrix` generates the communicability matrix.

source